+++ guillotina/interfaces/__init__.py
@@ -30,6 +30,7 @@ from .content import IResourceFactory  # noqa
 from .content import IStaticDirectory  # noqa
 from .content import IStaticFile  # noqa
 from .content import ITraversable  # noqa
+from .events import IBeforeFieldModified  # noqa
 from .events import IBeforeObjectAddedEvent  # noqa
 from .events import IBeforeObjectMovedEvent  # noqa
 from .events import IBeforeObjectRemovedEvent  # noqa