+++ CHANGELOG.rst
@@ -7,6 +7,7 @@
 - IBeforeFieldModified event to hook before setting a field
   [bloodbare]
 
+- Added PatchField for ints to be used as counters [lferran]
 
 4.1.3 (2018-08-08)
 ------------------
@@ -101,7 +102,7 @@
   [vangheem]
 
 - Dublin core should not be required
-  [bloodbare] 
+  [bloodbare]
 
 4.0.3 (2018-07-16)
 ------------------